+++ pact_methodology/carbon_footprint/characterization_factors.py
