+++ pact_methodology/carbon_footprint/product_or_sector_specific_rule_operator.py
