+++ pact_methodology/product_footprint/cpc.py
