+++ tests/carbon_footprint/test_carbon_footprint.py
@@ -1,31 +1,31 @@
 import pytest
 
-from pathfinder_framework.assurance.assurance import (
+from pact_methodology.assurance.assurance import (
     Assurance,
     Coverage,
     Level,
     Boundary,
 )
-from pathfinder_framework.carbon_footprint.carbon_footprint import CarbonFootprint
-from pathfinder_framework.carbon_footprint.characterization_factors import (
+from pact_methodology.carbon_footprint.carbon_footprint import CarbonFootprint
+from pact_methodology.carbon_footprint.characterization_factors import (
     CharacterizationFactors,
 )
-from pathfinder_framework.carbon_footprint.cross_sectoral_standard import (
+from pact_methodology.carbon_footprint.cross_sectoral_standard import (
     CrossSectoralStandard,
 )
-from pathfinder_framework.carbon_footprint.declared_unit import DeclaredUnit
-from pathfinder_framework.datetime import DateTime
-from pathfinder_framework.carbon_footprint.reference_period import ReferencePeriod
-from pathfinder_framework.carbon_footprint.geographical_scope import (
+from pact_methodology.carbon_footprint.declared_unit import DeclaredUnit
+from pact_methodology.datetime import DateTime
+from pact_methodology.carbon_footprint.reference_period import ReferencePeriod
+from pact_methodology.carbon_footprint.geographical_scope import (
     CarbonFootprintGeographicalScope,
     GeographicalGranularity,
 )
-from pathfinder_framework.data_quality_indicators.data_quality_indicators import (
+from pact_methodology.data_quality_indicators.data_quality_indicators import (
     DataQualityIndicators,
 )
-from pathfinder_framework.carbon_footprint.biogenic_accounting_methodology import BiogenicAccountingMethodology
-from pathfinder_framework.carbon_footprint.product_or_sector_specific_rule import ProductOrSectorSpecificRule
-from pathfinder_framework.carbon_footprint.product_or_sector_specific_rule_operator import ProductOrSectorSpecificRuleOperator
+from pact_methodology.carbon_footprint.biogenic_accounting_methodology import BiogenicAccountingMethodology
+from pact_methodology.carbon_footprint.product_or_sector_specific_rule import ProductOrSectorSpecificRule
+from pact_methodology.carbon_footprint.product_or_sector_specific_rule_operator import ProductOrSectorSpecificRuleOperator
 
 
 @pytest.fixture
