+++ tests/carbon_footprint/test_product_or_sector_specific_rule.py
@@ -1,9 +1,9 @@
 import pytest
 
-from pathfinder_framework.carbon_footprint.product_or_sector_specific_rule import (
+from pact_methodology.carbon_footprint.product_or_sector_specific_rule import (
     ProductOrSectorSpecificRule,
 )
-from pathfinder_framework.carbon_footprint.product_or_sector_specific_rule_operator import (
+from pact_methodology.carbon_footprint.product_or_sector_specific_rule_operator import (
     ProductOrSectorSpecificRuleOperator,
 )
