+++ tests/product_footprint/test_company_id_list.py
@@ -1,8 +1,8 @@
 import pytest
 
-from pathfinder_framework.exceptions import DuplicateIdError
-from pathfinder_framework.urn import CompanyId
-from pathfinder_framework.product_footprint.company_id_list import CompanyIdList
+from pact_methodology.exceptions import DuplicateIdError
+from pact_methodology.urn import CompanyId
+from pact_methodology.product_footprint.company_id_list import CompanyIdList
 
 
 def test_company_id_list_valid_company_ids():
